# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMPADU Keuangan API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/keuangan/jenis-tagihan": {
            "get": {
                "tags": [
                    "Jenis Tagihan"
                ],
                "summary": "Ambil semua jenis tagihan",
                "operationId": "67ab77997896fc009d5a4e3961ac62a2",
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/JenisTagihan"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Jenis Tagihan"
                ],
                "summary": "Tambah jenis tagihan baru",
                "operationId": "76c9f53a2758d9d8dd29eea287066200",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_jenis_tagihan"
                                ],
                                "properties": {
                                    "nama_jenis_tagihan": {
                                        "type": "string",
                                        "example": "UKT"
                                    },
                                    "deskripsi_tagihan": {
                                        "type": "string",
                                        "example": "Uang Kuliah Tunggal"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JenisTagihan"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/keuangan/jenis-tagihan/{id}": {
            "get": {
                "tags": [
                    "Jenis Tagihan"
                ],
                "summary": "Detail jenis tagihan",
                "operationId": "9875503586cabbc174823f6f91717610",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JenisTagihan"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Jenis Tagihan"
                ],
                "summary": "Update jenis tagihan",
                "operationId": "0ec6ebdcb3277cf737adab68e0b8f70a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_jenis_tagihan": {
                                        "type": "string"
                                    },
                                    "deskripsi_tagihan": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JenisTagihan"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Jenis Tagihan"
                ],
                "summary": "Hapus jenis tagihan",
                "operationId": "6fa6405417bc868f76652cf33112730c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "JenisTagihan": {
                "required": [
                    "nama_jenis_tagihan"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_jenis_tagihan": {
                        "type": "string",
                        "example": "UKT"
                    },
                    "deskripsi_tagihan": {
                        "type": "string",
                        "example": "Uang Kuliah Tunggal"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}